backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
SQLAlchemy==2.0.49
psycopg2-binary==2.9.12
python-dotenv==1.2.2
pytest==8.3.4
httpx==0.28.1
